my analysis file with your addition
I've re-loaded the analysis file, commenting out our for-loopy part and adding your recent addendum at the bottom, which runs beautifully. 

However, it looks like it generates a $count of 0 for every school in the data set, so I'm going to have to do some digging because that could mean that the data are crappy. I don't think this is the case, though, with 100km being the MAX_D at least SOME variation should exist.